+++ packages/tx-parser/src/services/bonkrewards/index.ts
@@ -1,19 +1,52 @@
 import { NetworkId } from '@sonarwatch/portfolio-core';
 import { ServiceDefinition } from '../../ServiceDefinition';
 
+const platformId = 'bonkrewards';
 const contract = {
   name: 'Staking',
   address: 'STAKEkKzbdeKkqzKpLkNQD3SUuLgshDKCD7U8duxAbB',
-  platformId: 'bonkrewards',
+  platformId,
 };
 
-const service: ServiceDefinition = {
-  id: 'bonkrewards-staking',
+const fireContract = {
+  name: 'Fire',
+  address: 'FiRESpaNzgYUiba5vkb44CZJLZjrux1AUECdfwPRsNkg',
+  platformId,
+};
+
+const swapContract = {
+  name: 'Swap',
+  address: 'BSwp6bEBihVLdqJRKGgzjcGLHkcTuzmSo1TQkHepzH8p',
+  platformId,
+};
+
+const stakingService: ServiceDefinition = {
+  id: `${platformId}-staking`,
   name: 'Staking',
-  platformId: 'bonkrewards',
+  platformId,
   networkId: NetworkId.solana,
   contracts: [contract],
 };
 
-export const services: ServiceDefinition[] = [service];
+const fireService: ServiceDefinition = {
+  id: `${platformId}-fire`,
+  name: 'Fire',
+  platformId,
+  networkId: NetworkId.solana,
+  contracts: [fireContract],
+};
+
+const swapService: ServiceDefinition = {
+  id: `${platformId}-swap`,
+  name: 'Swap',
+  platformId,
+  networkId: NetworkId.solana,
+  contracts: [swapContract],
+};
+
+export const services: ServiceDefinition[] = [
+  stakingService,
+  fireService,
+  swapService,
+];
 export default services;